# documentation/swagger-api.yaml
swagger: "2.0"
info:
  description: "EPRO Projekt WS2021"
  version: "1.0.0"
  title: "EPRO OKR Projekt"
host: "localhost:8080"
basePath: "/api/v1"
schemes:
  - "http"
paths:
...
  /businessUnit/{id}:
    get:
      tags:
        - "Business Unit"
      summary: "Get a Business Unit Object"
      produces:
        - "application/json"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of the Business Unit to return"
          required: true
          type: "integer"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/BusinessUnit"
    put:
      tags:
        - "Business Unit"
      summary: "Update a Business Unit"
      produces:
        - "application/json"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of the Business Unit to update"
          required: true
          type: "integer"
        - in: "body"
          name: "body"
          description: "The new Business Unit Object"
          required: true
          schema:
            $ref: "#/definitions/BusinessUnit"
      responses:
        200:
          description: "Successfully Updated the Business Unit"

    delete:
      tags:
        - "Business Unit"
      summary: "Delete a Business Unit"
      parameters:
        - name: "id"
          in: path
          description: ID of the Business Unit to be delted
          required: true
          type: integer
      responses:
        200:
          description: successfully deleted
        404:
          description: not found
...
  /businessUnitObjective/{objectiveId}:
    get:
      tags:
        - "Business Unit Objectives"
      summary: "Get the specified Business Unit objective"
      produces:
        - "application/json"
      parameters:
        - name: "objectiveId"
          in: "path"
          description: "Id of the Business Unit Objective"
          required: true
          type: "integer"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/BusinessUnitObjective"
    put:
      tags:
        - "Business Unit Objectives"
      summary: "Update the specified Business Unit objective"
      consumes:
        - "application/json"
      parameters:
        - name: "objectiveId"
          in: "path"
          description: "Id of the Business Unit Objective"
          required: true
          type: "integer"
        - in: "body"
          name: "body"
          description: "The updated Business Unit Objective"
          required: true
          schema:
            $ref: "#/definitions/BusinessUnitObjective"
      responses:
        200:
          description: "successful operation"
    delete:
      tags:
        - "Business Unit Objectives"
      summary: "delete the specified Business Unit objective"
      parameters:
        - name: "objectiveId"
          in: "path"
          description: "Id of the Business Unit Objective"
          required: true
          type: "integer"
      responses:
        200:
          description: "successful operation"
        404:
          description: "not found"
...
  /businessUnitObjective/{buoId}/keyResults:
    get:
      tags:
        - "Business Unit Key Results"
      summary: "Returns all Key Results of the specified Business Unit Objective"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          name: "buoId"
          description: "The BusinessUnitObjective for which the Keyresult are to be returned"
          required: true
          type: "integer"
      responses:
        200:
          description: "successfully got all BUO-KRs"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/BusinessUnitKeyResult"
    post:
      tags:
        - "Business Unit Key Results"
      summary: "Create a new Key Result for the specified Business Unit Objective"
      consumes:
        - "application/json"
      parameters:
        - in: "path"
          name: "buoId"
          description: "The Business Unit Objective for which the new Keyresult shall be created"
          required: true
          type: "integer"
        - in: "body"
          name: "body"
          description: "Key Result to be added"
          required: true
          schema:
            $ref: "#/definitions/BusinessUnitKeyResult"
      responses:
        200:
          description: "successfully created key result"
  /businessUnitKeyResults/{id}:
    get:
      tags:
        - "Business Unit Key Results"
      summary: "Return the specified Business Unit Key Result"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          name: "id"
          description: "The id of the BUO-KR to be returned"
          required: true
          type: "integer"
      responses:
        200:
          description: "Successfully returned keyresult"
          schema:
            $ref: "#/definitions/BusinessUnitKeyResult"
    put:
      tags:
        - "Business Unit Key Results"
      summary: "Update the specified Business Unit Key Result"
      consumes:
        - "application/json"
      parameters:
        - in: "path"
          name: "id"
          description: "The id of the BU-KR to be updated"
          required: true
          type: "integer"
        - in: "body"
          name: "body"
          description: "The updated Key Result Object"
          required: true
          schema:
            $ref: "#/definitions/BusinessUnitKeyResult"
      responses:
        200:
          description: "Successfully updated"
    delete:
      tags:
        - "Business Unit Key Results"
      summary: "Delete the specified Business Unit Key Result"
      consumes:
        - "application/json"
      parameters:
        - in: "path"
          name: "id"
          description: "The id of the BU-KR to be updated"
          required: true
          type: "integer"
      responses:
        200:
          description: "Successfully updated"
        404:
          description: "Successfully deleted"
  /businessUnitKeyResults/{id}/history:
    get:
      tags:
        - "Business Unit Key Results"
      summary: "Returns the history of the specified BusinessUnit Key Result"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          name: "id"
          description: "The id of the BU-KR of which the history is to be returned"
          required: true
          type: "integer"
      responses:
        200:
          description: "Successfully returned CompanyKeyResult History"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/BusinessUnitKeyResult"
...
definitions:
...
  BusinessUnit:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
  BusinessUnitObjective:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      achievement:
        type: "number"
      businessUnitId:
        type: "integer"
...
  BusinessUnitKeyResult:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      confidenceLvl:
        type: "number"
      comment:
        type: "string"
      goalValue:
        type: "number"
      currentValue:
        type: "number"
      achievement:
        type: "number"
      businessUnitObjectiveId:
        type: "integer"
...